Rename whitelist.js to origin-whitelist.js and sort imports.
