+++ extensions/amp-story/0.1/origin-whitelist.js
